# Conflicts:
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.infra/src/main/java/nts/uk/ctx/at/shared/infra/entity/workrule/statutoryworktime/flex/KshstFlxGetPrwkTime.java

+++ nts.uk/uk.com/com.web/nts.uk.com.web/src/main/webapp/view/ccg/034/d/ccg034.d.vm.ts
@@ -100,6 +100,7 @@ module nts.uk.com.view.ccg034.d {
         .mouseenter(() => vm.isMouseInsideLayout(true))
         .mouseleave(() => vm.isMouseInsideLayout(false))
         .mousedown((event) => {
+          $(".part-setting").filter((index, e) => $(e).css('display') === ('block')).removeAttr("style");
           // If layout in copy mode and mouse cursor is stay inside layout
           if (vm.isCopying() && vm.isMouseInsideLayout()) {
             // Stop copy mode
@@ -229,6 +230,7 @@ module nts.uk.com.view.ccg034.d {
     private createDOMFromData(partData: PartDataModel): JQuery {
       const vm = this;
       let $newPartTemplate = null;
+      let isInside = false;
       switch (partData.partType) {
         case MenuPartType.PART_MENU:
           $newPartTemplate = $("<div>", { "class": CSS_CLASS_MENU_CREATION_ITEM_CONTAINER }).append($('<div>', { 'class': `${CSS_CLASS_MENU_CREATION_ITEM} part-menu` }));
@@ -252,17 +254,22 @@ module nts.uk.com.view.ccg034.d {
           $newPartTemplate = $("<div>", { "class": CSS_CLASS_MENU_CREATION_ITEM_CONTAINER }).append($('<div>', { 'class': `${CSS_CLASS_MENU_CREATION_ITEM} part-menu` }));
           break;
       }
-      const $newPart: JQuery = LayoutUtils.renderPartDOM($newPartTemplate, partData);
+      const $newPart: JQuery = LayoutUtils.renderPartDOM($newPartTemplate, partData)
+        .hover(
+          handlerIn => LayoutUtils.onPartHover($newPart, true),
+          handlerOut => LayoutUtils.onPartHover($newPart, false))
+        .on("mousedown", event => LayoutUtils.onPartClickSetting($newPart, isInside));
       // Render div setting
-      let isInside = false;
       const $partSetting: JQuery = $("<div>", { "class": 'part-setting' })
         .hover(
           (handlerIn) => isInside = true,
-          (handlerOut) => {
-            isInside = false;
-            console.log(isInside);
-          })
+          (handlerOut) => isInside = false)
         .on("click", event => {
+          event.stopPropagation();
+          // Ignore when clicked inside the popup option
+          if (event.target.classList.contains("part-setting-popup-option")) {
+            isInside = false;
+          }
           LayoutUtils.onPartClickSetting($newPart, isInside);
         });
       const $partSettingPopup: JQuery = $("<div>", { "class": 'part-setting-popup' })
@@ -281,10 +288,7 @@ module nts.uk.com.view.ccg034.d {
           .on('click', (event) => {
             LayoutUtils.onPartClickSetting($newPart, false);
             vm.removePart($newPart);
-          }))
-        .hover(
-          (handlerIn) => isInside = true,
-          (handlerOut) => isInside = false);
+          }));
       $partSettingPopup.appendTo($partSetting);
       $partSetting.appendTo($newPart);
       // Check and remove overlap part (both DOM element and data by calling JQuery.remove())
@@ -1004,14 +1008,12 @@ module nts.uk.com.view.ccg034.d {
                 vm.mapPartData[partClientId] = result.partData;
                 // Update part DOM
                 LayoutUtils.renderPartDOMAttachment($part, result.partData as PartDataAttachmentModel);
-              } else if (isCreateDialog) {
-                // If this is dialog setitng when create => remove part
-                vm.removePart($part);
-                if (result.fileId) {
-                  vm.removeFile(result.fileId);
+              } else {
+                if (isCreateDialog) {
+                  // If this is dialog setitng when create => remove part
+                  vm.removePart($part);
                 }
-              } else if (result.fileId) {
-                vm.modifiedPartList.added.push(result.fileId);
+                vm.removeFile(result.fileId);
               }
             });
           break;
@@ -1020,27 +1022,23 @@ module nts.uk.com.view.ccg034.d {
             .then((result: any) => {
               if (result.isSaving) {
                 const partImage = (result.partData as PartDataImageModel);
-                if (partImage.isFixed === 1 && partImage.fileId) {
-                  // Only prepare for delete if has changes
-                  if (!_.find(vm.modifiedPartList.added, partImage.fileId) && partImage.fileId !== partImage.originalFileId) {
-                    vm.modifiedPartList.added.push(partImage.fileId);
-                  }
-                  if (!_.find(vm.modifiedPartList.deleted, partImage.originalFileId) && partImage.fileId !== partImage.originalFileId) {
-                    vm.modifiedPartList.deleted.push(partImage.originalFileId);
-                  }
+                // Only prepare for delete if has changes
+                if (!_.find(vm.modifiedPartList.added, partImage.fileId) && partImage.fileId !== partImage.originalFileId) {
+                  vm.modifiedPartList.added.push(partImage.fileId);
+                }
+                if (!_.find(vm.modifiedPartList.deleted, partImage.originalFileId) && partImage.fileId !== partImage.originalFileId) {
+                  vm.modifiedPartList.deleted.push(partImage.originalFileId);
                 }
                 // Update part data
                 vm.mapPartData[partClientId] = result.partData;
                 // Update part DOM
                 LayoutUtils.renderPartDOMImage($part, result.partData as PartDataImageModel);
-              } else if (isCreateDialog) {
-                // If this is dialog setitng when create => remove part
-                vm.removePart($part);
-                if (result.fileId) {
-                  vm.removeFile(result.fileId);
+              } else {
+                if (isCreateDialog) {
+                  // If this is dialog setitng when create => remove part
+                  vm.removePart($part);
                 }
-              } else if (result.fileId) {
-                vm.modifiedPartList.added.push(result.fileId);
+                vm.removeFile(result.fileId);
               }
             });
           break;
@@ -1336,6 +1334,9 @@ module nts.uk.com.view.ccg034.d {
       htmlContent += ` xmlns:com="http://xmlns.jcp.org/jsf/component"`;
       htmlContent += ` xmlns:com="http://xmlns.jcp.org/jsf/html"`;
       htmlContent += `<head>`;
+      htmlContent += `<style>`;
+      htmlContent += `.ccg034-component-container { word-break: break-all; word-break: break-word; word-wrap: break-word }`;
+      htmlContent += `</style>`;
       htmlContent += `<link rel="stylesheet" type="text/css" href="/nts.uk.com.js.web/lib/nittsu/ui/style/stylesheets/base.css">`;
       htmlContent += `</head>`;
       htmlContent += `<body>`;
@@ -1396,7 +1397,9 @@ module nts.uk.com.view.ccg034.d {
      * 変更前のファイルは削除
      */
     private removeFile(fileId: string) {
-      (nts.uk.request as any).file.remove(fileId);
+      if (!nts.uk.text.isNullOrEmpty(fileId)) {
+        (nts.uk.request as any).file.remove(fileId);
+      }
     }
 
     private updateOriginalFileId() {
@@ -1435,12 +1438,28 @@ module nts.uk.com.view.ccg034.d {
      * @param partClientId
      */
     static onPartClickSetting($part: JQuery, visible: boolean) {
+      const $partSetting: JQuery = $part.find('.part-setting');
       const $partSettingPopup: JQuery = $part.find('.part-setting-popup');
       if ($partSettingPopup) {
         if (visible) {
           $partSettingPopup.css('display', 'block');
+          $partSetting.css('display', 'block');
         } else {
           $partSettingPopup.css('display', 'none');
+        }
+      }
+    }
+
+    static onPartHover($part: JQuery, isHoverIn: boolean) {
+      const $partSetting: JQuery = $part.find('.part-setting');
+      const $partSettingPopup: JQuery = $part.find('.part-setting-popup');
+      if ($partSettingPopup.css("display") !== "none") {
+        $partSetting.css('style');
+      } else {
+        if (isHoverIn) {
+          $partSetting.css('display', 'block');
+        } else {
+          $partSetting.css('display', 'none');
         }
       }
     }
@@ -1853,7 +1872,7 @@ module nts.uk.com.view.ccg034.d {
               'cursor': 'pointer',
             });
           $partHTML = $("<div>")
-            .attr('style', '-ms-word-break: break-all')
+            .addClass("ccg034-component-container")
             .css({
               'position': 'absolute',
               'top': `${partDataMenuModel.positionTop}px`,
@@ -1865,7 +1884,6 @@ module nts.uk.com.view.ccg034.d {
               'justify-content': LayoutUtils.getHorizontalClass(partDataMenuModel.alignHorizontal),
               'overflow': 'hidden',
               'text-overflow': 'ellipsis',
-              'word-break': 'break-word',
             })
             .append($partMenuHTML);
           break;
@@ -1880,7 +1898,7 @@ module nts.uk.com.view.ccg034.d {
               'white-space': 'pre-wrap',
             });
           $partHTML = $("<div>")
-            .attr('style', '-ms-word-break: break-all')
+            .addClass("ccg034-component-container")
             .css({
               'position': 'absolute',
               'top': `${partDataLabelModel.positionTop}px`,
@@ -1892,7 +1910,6 @@ module nts.uk.com.view.ccg034.d {
               'justify-content': LayoutUtils.getHorizontalClass(partDataLabelModel.alignHorizontal),
               'overflow': 'hidden',
               'text-overflow': 'ellipsis',
-              'word-break': 'break-word',
               'background-color': partDataLabelModel.backgroundColor,
             })
             .append($partLabelHTML);
@@ -1910,7 +1927,7 @@ module nts.uk.com.view.ccg034.d {
               'cursor': 'pointer',
             });
           $partHTML = $("<div>")
-            .attr('style', '-ms-word-break: break-all')
+            .addClass("ccg034-component-container")
             .css({
               'position': 'absolute',
               'top': `${partDataLinkModel.positionTop}px`,
@@ -1922,7 +1939,6 @@ module nts.uk.com.view.ccg034.d {
               'justify-content': LayoutUtils.getHorizontalClass(partDataLinkModel.alignHorizontal),
               'overflow': 'hidden',
               'text-overflow': 'ellipsis',
-              'word-break': 'break-word',
             })
             .append($partLinkHTML);
           break;
@@ -1940,7 +1956,7 @@ module nts.uk.com.view.ccg034.d {
               'cursor': 'pointer',
             });
           $partHTML = $("<div>")
-            .attr('style', '-ms-word-break: break-all')
+            .addClass("ccg034-component-container")
             .css({
               'position': 'absolute',
               'top': `${partDataAttachmentModel.positionTop}px`,
@@ -1952,7 +1968,6 @@ module nts.uk.com.view.ccg034.d {
               'justify-content': LayoutUtils.getHorizontalClass(partDataAttachmentModel.alignHorizontal),
               'overflow': 'hidden',
               'text-overflow': 'ellipsis',
-              'word-break': 'break-word',
             })
             .append($partAttachmentHTML);
           break;
